Revert "complilations Errors resolved of repo"
This reverts commit d04bc716d1f90ac332c9c5808d5548f740255a0b.

+++ src/main/java/com/iemr/mmu/repo/masterrepo/doctor/PreMalignantLesionMasterRepo.java
@@ -29,7 +29,7 @@ import org.springframework.stereotype.Repository;
 
 import com.iemr.mmu.data.masterdata.doctor.PreMalignantLesion;
 @Repository
-public interface PreMalignantLesionMasterRepo extends CrudRepository<PreMalignantLesion, Integer>{
+public interface PreMalignantLesionMasterRepo extends CrudRepository<PreMalignantLesion, Long>{
 
 	@Query("select preMalignantLesionID, preMalignantLesionType from PreMalignantLesion where deleted = false "
 			+ "order by preMalignantLesionType")
