+++ src/main/java/com/iemr/mmu/repo/masterrepo/doctor/V_DrugPrescriptionRepo.java
@@ -31,7 +31,7 @@ import org.springframework.stereotype.Repository;
 import com.iemr.mmu.data.masterdata.doctor.V_DrugPrescription;
 
 @Repository
-public interface V_DrugPrescriptionRepo extends CrudRepository<V_DrugPrescription, Integer> {
+public interface V_DrugPrescriptionRepo extends CrudRepository<V_DrugPrescription, Long> {
 	@Query("SELECT t FROM V_DrugPrescription t WHERE t.facilityID =:facilityID ")
 	public ArrayList<V_DrugPrescription> getItemListForFacility(@Param("facilityID") Integer facilityID);
 }
