+++ src/main/java/com/iemr/mmu/repo/masterrepo/ncdScreening/NCDScreeningReasonRepo.java
@@ -30,7 +30,7 @@ import org.springframework.stereotype.Repository;
 import com.iemr.mmu.data.masterdata.ncdscreening.NCDScreeningReason;
 
 @Repository
-public interface NCDScreeningReasonRepo extends CrudRepository<NCDScreeningReason, Integer>{
+public interface NCDScreeningReasonRepo extends CrudRepository<NCDScreeningReason, Long>{
 
 	@Query("select ncdScreeningReasonID, ncdScreeningReason from NCDScreeningReason where deleted = false order by ncdScreeningReason")
 	public ArrayList<Object[]> getNCDScreeningReasons();
